fix: use mock data in lesson overview video stories
