+++ src/components/TeacherComponents/LessonOverviewVideo/LessonOverviewVideo.fixtures.ts
@@ -1,0 +1,21 @@
+import { AnalyticsBrowseData } from "@/components/TeacherComponents/types/lesson.types";
+
+export const mockBrowsePathwayData: AnalyticsBrowseData = {
+  keyStageSlug: "ks4",
+  keyStageTitle: "Key stage 4",
+  subjectSlug: "maths",
+  subjectTitle: "Maths",
+  unitSlug: "surds",
+  unitName: "Surds",
+  lessonSlug: "lesson-1",
+  lessonName: "Lesson 1",
+  pathway: null,
+  tierName: "Foundation",
+  yearGroupName: "Year 10",
+  yearGroupSlug: "year-10",
+  examBoard: null,
+  releaseGroup: "2023",
+  phase: "secondary",
+  lessonReleaseCohort: "2023-2026",
+  lessonReleaseDate: "unreleased",
+};
