+++ src/components/TeacherComponents/LessonOverviewVideo/LessonOverviewVideo.test.tsx
@@ -1,11 +1,10 @@
 import { fireEvent } from "@testing-library/react";
 
-import { AnalyticsBrowseData } from "../types/lesson.types";
-
 import {
   LessonOverviewVideo,
   LessonOverviewVideoProps,
 } from "./LessonOverviewVideo";
+import { mockBrowsePathwayData } from "./LessonOverviewVideo.fixtures";
 
 import renderWithTheme from "@/__tests__/__helpers__/renderWithTheme";
 
@@ -13,26 +12,6 @@ jest.mock("@/components/SharedComponents/VideoPlayer/VideoPlayer", () => ({
   __esModule: true,
   default: () => <video data-testid="video-element" />,
 }));
-
-export const mockBrowsePathwayData: AnalyticsBrowseData = {
-  keyStageSlug: "ks4",
-  keyStageTitle: "Key stage 4",
-  subjectSlug: "maths",
-  subjectTitle: "Maths",
-  unitSlug: "surds",
-  unitName: "Surds",
-  lessonSlug: "lesson-1",
-  lessonName: "Lesson 1",
-  pathway: null,
-  tierName: "Foundation",
-  yearGroupName: "Year 10",
-  yearGroupSlug: "year-10",
-  examBoard: null,
-  releaseGroup: "2023",
-  phase: "secondary",
-  lessonReleaseCohort: "2023-2026",
-  lessonReleaseDate: "unreleased",
-};
 
 describe("LessonOverviewVideo", () => {
   it("Renders the video player", () => {
